refactor(timesheets): move RecurringManager + edit modal into timesheet/
Both files were sitting at the components/ root instead of in the
timesheet module folder alongside DailyView and WeeklyView. Move them
into components/timesheet/ to match codebase convention. Update relative
imports (../ -> ../../, ./shared/ -> ../shared/) and the App.tsx
import path.
